+++ apache2/msc_util.c
@@ -274,7 +274,7 @@ unsigned char is_netmask_v6(char *ip_strv6) {
  *
  * \retval string On Success
  */
-char *parse_pm_content(const char *op_parm, unsigned short int op_len, msre_rule *rule, char **error_msg)  {
+char *parse_pm_content(const char *op_parm, size_t op_len, msre_rule *rule, char **error_msg)  {
     char *parm = NULL;
     char *content = NULL;
     unsigned short int offset = 0;
@@ -708,7 +708,7 @@ char *file_basename(apr_pool_t *mp, const char *filename) {
 
 char *m_strcasestr(const char *haystack, const char *needle) {
     char aux, lower_aux;
-    int length;
+    size_t length;
 
     if ((aux = *needle++) != 0) {
         aux = (char)tolower((unsigned char)aux);
